+++ CHANGELOG.md
@@ -8,8 +8,20 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Changed
 
+- `slider`
+  - An upcoming change will migrate the slider to use the MDC M2 slider. In
+    preparation for this, the MWC slider => MDC slider dependency will not be
+    updated until the migration is complete.
 - `top-app-bar(-fixed)`
   - `--mdc-top-app-bar-width` is now configurable
+- `mwc-fab`
+  - **BREAKING** removed `--mdc-fab-box-shadow-hover`; use `--mdc-fab-box-shadow`
+  - **BREAKING** removed `--mdc-fab-box-shadow-active`; use `--mdc-fab-box-shadow`
+  - Ripple now uses and exposes `mwc-ripple`'s CSS custom properties API
+- `mwc-icon-button`
+  - **BREAKING** now uses lazy `mwc-ripple` in its implementation
+- `mwc-icon-button-toggle`
+  - **BREAKING** now uses lazy `mwc-ripple` in its implementation
 
 ### Fixed
 
@@ -22,7 +34,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - prod dependency breakage based on `@material/dom`
 
-## [v0.17.2] - 2020-01-01
+## [v0.17.1] - 2020-01-01
 
 - unused GH tag